set code as legacy
