+++ app/Actions/Import/FromServer.php
@@ -2,7 +2,7 @@
 
 namespace App\Actions\Import;
 
-use App\Actions\Photo\Strategies\ImportMode;
+use App\DTO\ImportMode;
 use App\Models\Album;
 use function Safe\ini_get;
 use Symfony\Component\HttpFoundation\StreamedResponse;
